// src/models/mst_mandatory_hdmf.ts
class mst_mandatory_hdmf{
    id?: number = 0;
    amount_start: number = 0;
    amount_end: number = 0;
    employee_contribution_percentage: number = 0;
    employer_contribution_percentage: number = 0;
    employee_contribution_value: number = 0;
    employer_contribution_value: number = 0;
    remarks: string = "";

    
  }
  export default mst_mandatory_hdmf;